requirements.txt
dashscope>=1.14.0
zhipuai>=2.0.0
tiktoken>=0.5.0
pyyaml>=6.0
rich>=13.0.0
python-dotenv>=1.0.0